Add logging conditions similar to production + add tests to repo_providers
We are having issues that the torngit tests are not catching: Under normal circunstances, the logging functions accept any parameter. But with our
logging structure, they all need to be inside an extra dict. To catch this, we are adding an initial configure step so we can see the tests reflect those logging errors