Implementing fixes to fix CommentNotifier usage
- Upgrading requirements.txt
- Adding check in case yaml comments is False
- Removing plus sign in case value is zero

There was a small oversight at my end regarding get_config("site", "comment"). Due to the purpose of the "site" configuration value, the value of get_config("site", "comment") is already the default of the current 'yaml->comment'. So there is no point in using the original value: it is either already in the yaml, or the user overwrote and the right behavior.