Making UserGivenSecret return clean value
In the past, when we didnt know how we would use secrets, we left the function returning the whole thing.

But now some users really use it, so we need to return clean values so the fields reading them can use them.

In an ideal situation we would validate agains the user owner and repo data. But that will only be needed when we expose this on /validate