Forcing change of content_type on GCP
In some cases, it seems like our bash uploader (and possibly other uploaders) upload a file using

s3=

which is an improper Content-Type given the type of the file, as documented in https://cloud.google.com/storage/docs/transcoding\#content-type_vs_content-encoding

Unfortunately, enforcing the clients to use the proper header might be a problem, since some users have downloaded the uploader and will keep using an old version of them.

So the quickest way to get this working is to manually force the content-types to text/plain on download time.

While this is not great and might be extra API calls, we should also move to update the client uploaders so at least they dont have this issue