Small experiment on what rolling out own cache would do
Here is how cache is meant to be used. Let's say you have a function f (sync or async). You can then simply do

from helpers.cache import cache

@cache.cache_function
def f(...):
    ....

And it is done.

The way this is currently done is by using Redis as backend and pickling and unpickling the cached result as needed. The cache works per-parameter, so the there are no side-effects on other calls of the function.

The function parameters does need to have a sensible __repr__. Notice that this means the original meaning of __repr__: a string that is unique to that instance, and is consistently that value for the relevant parameters (ie, you dont want to use time.now() as part of the repr, because althought this makes the instance unique, on other occasions this instance will not be this same value and the cache will not be properly reused).

The TTL of the cache is for now pre-determined globally. We can tune as we need, but I dont want us to overengineer until we properly understand how much this cache is helping us