Making it so publioc repos use the public bots we have in place
For some reason we dont yet understand, public repos are getting set with the wrong repo.owner.bot. It is unfortunate, and users see this as a big issue (correctly).

The temp solution is to implement what the legacy comment had, but more globally: public repos will now use the public bots we have assigned to them. But now this will be the widespread behavior instead of only comment-wise.

We still hve to figure out why the bots are being set wrong on the owner.bot. But this solution should stop the unwanted comments for now