Not using base-aware pathfixing sometimes
In cases where the toc (list of network files) are not provided by the user, using the base path will always lead to the final ReportFile.name be something like:

'www/repos/platform_project/reader/reader-app/coverage_c64c67ea/total'

Which is not that helpful.

Also trying to reduce the amount of logging that we do. It is too noisy and hard to track