Stopping race condition between Upload and UploadProcessor
We have recently had a problem with a customer using carryforward flags where, when picking a parent commit to carry the flags from, the system just says 'I cant find a parent commit'

Checking on the database always show the value on the 'parent' column on the db is correct. So it was either later filled or something in the system couldnt read it

Our current hypothesis is a race-condition. There is some time between the 'Upload' task calling 'self.schedule_task(commit, commit_yaml, argument_list)' and it actually calling 'db_session.commit()'. So we believe that if the UploadProcessor task starts and fetchs commit info in that meantime, it gets a version of that commit without the parent_commit_id

This probably wasnt relevant until now because the parent was never needed when processing reports before.

So our fix is going to be calling db_session.commit() before scheduling the other tasks. This is not an unprecedented change, as 'UploadProcessorTask' already does it to avoid having the following chained task fetch an outdated version of the report.

We dont know if that is the only issue causing that, but even if it isnt, the race condition issue is theorietically real, and there is no downside to calling db_session.commit earlier