Replacing astuple by instance methods
We recently found that dataclasses.astuple is a really slow method. The reason is that it does deep copies of all fields to generate the tuple.

While this is not usually a big deal, since we do that multiple times in our system inside reports (at least one per line when storing or exposing it to an API), it adds up to a lot.

So we are writing our own `def astuple(self)` methods. We dont have the requirement of deep copies, and we know exactly how the final object will be generated (we dont need the system to do meta-analysis on the class).