Dealing with cases where pull request is brand new
Apparently in the legacy code, whenever a pull is inserted in the db, it is inserted with head=NULL. The system will eventually populate it using the `commits_update_heads` trigger.

We are unsure whether that is a good idea. But in the spirit of porting it first, and fixing it later, we will try to apply this.

Also adding checks for when the pull has no head (like in the above case), the system should not attempt to dothe remaining of the syncing/notifying logic.