Finding a decent solution to pullsync race condition
We are still having some race condition issues. So we will make the proper fix: 'It is easier to ask for forgiveness than permission'

We will then try to insert the pull request first, and if that fails we will rollback the nested transaction and update an existing pull. This ensures race conditions between reading and writing dont happen.

In addition to make the proper order, it becomes easier to test, since that situation is easier to reproduce