Creating logic for determining if current system is running under enterprise
We have to find a way to determing if the current system is enterprise or not.

This is the way legacy does it. This, combined with a pyinstaller packaging, makes it not trivial for the users to just remove the file.

But we think that is not that hard, and we should develop better ways of doing it in the future, like adding special secrets that only production know (essentially making de-enterprising the package a matter of knowing a special passcode, not just removing a file).

But for now, this is no worse than legacy