Adding compatibility with different redis jobs key
We will soon make the new-worker the default worker for upload processing, and remove all code from legacy that conditionally sends tasks to different workers

The problem is that, in order to do the testing, we had to use different redis keys for the upload jobs. One for legacy and one for new.

Now, during the switch, we will have some jobs going to the new-style location, and some jobs going to the legacy-style location.

So we need the new-worker to accomodate and deal with both.

This is the way we will do it.

It's temporary only, so one day we hope to remove the 'testuplaods...' key-style.