Catching SoftTimeLimit exception during dbsession commit
There is this unusual situation happening on production sometimes

```
Traceback (most recent call last):
  File "/usr/local/lib/python3.7/site-packages/celery/app/trace.py", line 385, in trace_task
    R = retval = fun(*args, **kwargs)
  File "/usr/local/lib/python3.7/site-packages/celery/app/trace.py", line 648, in __protected_call__
    return self.run(*args, **kwargs)
  File "/usr/local/lib/python3.7/site-packages/sentry_sdk/integrations/celery.py", line 161, in _inner
    reraise(*exc_info)
  File "/usr/local/lib/python3.7/site-packages/sentry_sdk/_compat.py", line 57, in reraise
    raise value
  File "/usr/local/lib/python3.7/site-packages/sentry_sdk/integrations/celery.py", line 156, in _inner
    return f(*args, **kwargs)
  File "/worker/tasks/base.py", line 52, in run
    db_session.commit()
  File "/usr/local/lib/python3.7/site-packages/sqlalchemy/orm/session.py", line 954, in commit
    self.transaction.commit()
  File "/usr/local/lib/python3.7/site-packages/sqlalchemy/orm/session.py", line 477, in commit
    self._remove_snapshot()
  File "/usr/local/lib/python3.7/site-packages/sqlalchemy/orm/session.py", line 376, in _remove_snapshot
    s._expire(s.dict, self.session.identity_map._modified)
  File "/usr/local/lib/python3.7/site-packages/sqlalchemy/orm/state.py", line 571, in _expire
    del dict_[key]
  File "/usr/local/lib/python3.7/site-packages/billiard/pool.py", line 229, in soft_timeout_sighandler
    raise SoftTimeLimitExceeded()
billiard.exceptions.SoftTimeLimitExceeded: SoftTimeLimitExceeded()
```

where a SoftTimeLimitExceeded happens exactly inside the db_session.commit() code.

We are still unsure on how to fix it, but we know that this essentially breaks the session, since it puts it in an inconsistent state

Future calls on the same dbsession raise:

```
sqlalchemy.exc.InvalidRequestError: This session is in 'committed' state; no further SQL can be emitted within this transaction.
```

So we are trying to catch it and do some cleanup, because since the dbsession is reused, future tasks are also affected by this issue

We are unsure of whether this will work or not. But since this is not a normal occurrence, we hope we can adjust it as time goes by