Redoing flush repo
We had ported flush repo a long time ago (one of the first tasks we did)

Now that we know better and have better tooling and standards of coding, it makes sense to redo it with proper testing. Specially since it is a short task.

The flush repo task only deletes the repository information.