Syncing pull post-processing
Due to the fact that we recently removed pulls.sync task from the Upload flow, now there are cases where `pull.flare` and `pull.diff` are not being updated properly, because no webhooks come from github after the proessing is done.

To fix that we need to re-enable some sort of pull.sync scheduling when the upload processing is over

But we need to make sure thie pull.sync does not uncessarily calls notify again

While this is semantically correct (the commit is the head of the pull and the commit was updated, therefore do the necessary adjustments to the pull), this unfortunately feels al ittle clunky, since we mostly need the section of PullSyncTask that updates the pull.flare and pull.diff. Specially since this requires an ugly `if should_send_notifications` on the PullSyncTask.

But for now this might be the safest way to do it, since we are not sure how future developments will affect the syncing of pull requests.