Adding default hash key to match legacy behavior
In legacy, the minio hash_key (a key we use to obfuscate the path of the files uploaded by the customers) has a default.

We should add that default here as well to make for a smoother transition from legacy to new.

In production we already have that field set in the envvars (to a different value), so we should not need to change anything there