Attempting a different HTTP client
We seem to be hitting connection issues with Gitlab. They claim we are not properly closing the connection when we should.

To rule out a bad client, we are testing a different implementation og the http client that is supposedely more resilient to 'non-common' http features.

In the worst case, things stay like they are.

In the best case, Gitlab is fixed.

In the worst worst case, we break Github. But we will so exensive testing locally before to make sure this doesn't happen