Switching implementation of AsyncHTTPClient
We for some reason have issues on Gitlab where they seem to be closing the connection and we our client crashes after that

We think that switching the client implementation helps. We dont know why. But it looks like it works on local while the old implementation doesnt

But we need to install 'pycurl' for that. It requires some extra packages installed