Attempting new way of calculating Gitlab PR base
It looks like somewhere after Gitlab Base Picking was originally implemented, Gitlab added some fields to better determine pull base in the PR api response

This field is called `diff_refs` and it was documented art https://gitlab.com/gitlab-org/gitlab-foss/-/merge_requests/21923 (1 year ago), so it looks like its implementation is relatively recent.

We will try to use it, and also leave the current logic as fallback for the cases this information might not be there sometimes. We will also log the cases when the fallback is used, so we can understand such situations and why such code sometimes leave the same commit as base and head of a PR.