Fixing some case in CF we are not sure why the CFF is lost
Trying to fix a case where one of our clients, Compass, is losing their CFF somewhere along the way.

We are trying to better wrap and log the graph traversing we do when looking for a commit to CFF from, because we believe this might be related to that issue.

We also are considering that commit.state == 'skipped' is a valid thing to carryforward from, because that state only relates to notifications, not processing.
But since 'skipped' only happens when the commit matches the regex:

```
re.compile(r"\[(ci|skip| |-){3,}\]")
```

we dont think that's the cause.

Also trying to get the root cause of the situation for why the logs dont reach datadog. We believe it happens due to the size of this client YAML, making the log message be too big for datadog.