Adding logging to Github Integration obtaining
We have a situation where Github, under integrations, is returning a 401 HTTP error to us on a client (called Compass).

This 401 is very rare for that customer (3 times in the last week), and they say there's nothing unusual on their end on the situations the 401 was raised

So we dont have many ideas on what could the root cause be.

One of the remaining hypothesis is that the access_token we generated was expired. The expiration can be done manually (at least via API) or by time (we request access tokens that live 500 seconds).

We have yet to ask if the customer could have done something to somehow invalidate the access tokens by hand. But we are going to log some information in the access_token generation to make sure we are not missing something here.

In theory, it is extremely unlikely that we hit a time issue. The tokens are made to last 500 seconds. We cache them for 120 seconds. We noticed from start of task to the moment of failure it took 2 seconds. So in the worst case (where we were using a cached token close to expire), we have a token that lived 122 seconds.

There are things that could affect that 122s number:

- Github takes too long to respond with token: it took 80 seconds in a small window of time Wednesday to receive the token
- Redis could have failed to expire the token after 120 seconds (how much extra long can they take?)
- We made requests that took too long to reach Github andby then the token was expired?

So it doesn't seem likely that a 500 second limit was hit with the issues above. Also, the task where the failure occured seems to have made API calls with good responses after this call failed. So a token expiration alone doesnt seem enough to explain this.

 We are hoping to log the response json (sans the token) so we can find which of our assumptions is wrong.

It would be useful to also log part of the token in a way it's not a secutiry issue but could still tell us if Github was somehow reusing/caching tokens themselves. Because maybe their cache in combination with our cache could be part of a problem taht could cause those uncommon clashes