Adding CorruptRawReportError
Adding a special exception for when a report has some inconsistencies with its expected format.

Notice that this exception is not meant for 'Let's try this file on every processor, and just have the processor raise 'CorruptRawReportError' when it notices it does not match them. For high-level analysis of the format, we still should use `matches_content` logic. This exception is meant for cases where something deep inside the file is problematic, and it wouldnt be possible to just catch it unless `matches_content` reimplemented the whole processing of the file again.

Also implementing its usage on 'VOneProcessor'. That is a reasonable example of the inconsistency of the file. Notice that, if customers with 100s of files per report upload this, detecting its problems on 'matches_content' would require it to go through the whole file, and then one more pass through the whole file to actually process it.