Fixing case where config goes deeper than default config
In some cases, like Lyfts yaml, the user sets some fields that go more nested than the default config. For example, the default config has

```
"site": {
    "coverage": {
        "status": {"project": True, "patch": True, "changes": False}
    }
}
```

and the user has

```
site:
  coverage:
    status:
      project:
        default:
          only_pulls: true
          target: auto
          threshold: 100%
```

Our system doesnt know how to merge those configurations. We need to tell it that if the default is a non-dict and the user-provided is a dict then we should use the provided-dict.

So we are fixing that